Analysed the campaign data available and build a classification model without calls data
During any campaign, it is important to target the right customers. In this case its telemarketing campaign and if the customer doesn't want the product than the call would be ineffective. So it's best to classify the customer beforehand and call that customer who has a higher propensity to convert this will save marketing team time, calls, and money.